[add] timer 기능 play and pause 기능
