+++ Skins/Dali/vanila/doc/03_Timer.md
@@ -1,0 +1,11 @@
+### Timer
+
+
+
+```js
+this.musicTimer = new Timer({timer, audioEl: this.audioEl})
+이 부분 직접 관계를 안 맺고 하는 방법 
+timer는 지금 시간만 알고 display를 해준다. 
+audioEl먼트에 존재를 알 필요가 있을까? 
+
+```
